+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-## Wednesday, June 12th, 2024
+## Tuesday, June 18th, 2024
 
 ### Added
 
 * Display CTR percentages in the table when Looker is configured
+
+## Thursday, June 13th, 2024
+
+### Added
+
+* Added an option to make experiment tables default to being fully expanded or collapsed
 
 ## Monday, May 27th, 2024